+++ services/federator/src/Federator/Validation.hs
@@ -19,10 +19,12 @@ module Federator.Validation
   ( federateWith,
     validateDomain,
     throwInward,
+    sanitizePath,
   )
 where
 
 import Control.Lens (view)
+import qualified Data.ByteString as BS
 import Data.Domain (Domain, domainText, mkDomain)
 import Data.String.Conversions (cs)
 import Federator.Options
@@ -30,6 +32,7 @@ import Imports
 import Polysemy (Members, Sem)
 import qualified Polysemy.Error as Polysemy
 import qualified Polysemy.Reader as Polysemy
+import URI.ByteString
 import Wire.API.Federation.GRPC.Types
 
 -- | Validates an already-parsed domain against the allowList using the federator
@@ -60,3 +63,48 @@ validateDomain unparsedDomain = do
 
 throwInward :: Members '[Polysemy.Error InwardError] r => InwardErrorType -> Text -> Sem r a
 throwInward errType errMsg = Polysemy.throw $ InwardError errType errMsg
+
+-- | Normalize the path, and ensure the path begins with "federation/" after normalization
+sanitizePath :: Members '[Polysemy.Error InwardError] r => ByteString -> Sem r ByteString
+sanitizePath originalPath = do
+  when (BS.length originalPath > 200) $
+    throwInward IForbiddenEndpoint "path too long"
+  -- we parse the path using the URI.ByteString module to make use of its normalization functions
+  uriRef <- case parseRelativeRef strictURIParserOptions originalPath of
+    Left err -> throwInward IForbiddenEndpoint (cs $ show err <> cs originalPath)
+    Right ref -> pure ref
+
+  -- we don't expect any query parameters or other URL parts other than a plain path
+  when (queryPairs (rrQuery uriRef) /= []) $
+    throwInward IForbiddenEndpoint "query parameters not allowed"
+  when (isJust (rrFragment uriRef)) $
+    throwInward IForbiddenEndpoint "fragments not allowed"
+  when (isJust (rrAuthority uriRef)) $
+    throwInward IForbiddenEndpoint "authority not allowed"
+
+  -- Perform these normalizations:
+  -- - hTtP -> http
+  -- - eXaMpLe.org -> example.org
+  -- - If the scheme is known and the port is the default (e.g. 80 for http) it is removed.
+  -- - If the path is empty, set it to /
+  -- - Rewrite path from /foo//bar///baz to /foo/bar/baz
+  -- - Sorts parameters by parameter name
+  -- - Remove dot segments as per RFC3986 Section 5.2.4
+  let normalizedURI = normalizeURIRef' aggressiveNormalization uriRef
+  -- sometimes normalization above results in a path with a leading slash.
+  -- For consistency, remove a leading slash, if any
+  let withoutLeadingSlash = BS.stripPrefix "/" normalizedURI
+  let normalized = fromMaybe normalizedURI withoutLeadingSlash
+
+  -- to guard against double percent encoding, we disallow the '%' character
+  -- here, since we expect to only use POST requests on ASCII paths without any
+  -- query parameters
+  let (_, pTailEncoding) = BS.breakSubstring "%" normalized
+  unless (BS.null pTailEncoding) $
+    throwInward IForbiddenEndpoint "percent encoding not allowed"
+
+  -- Most importantly, only allow paths with the federation/ prefix.
+  unless ("federation/" `BS.isPrefixOf` normalized) $
+    throwInward IForbiddenEndpoint ("disallowed path: " <> cs originalPath)
+
+  pure normalized