+++ services/spar/test-integration/Test/Spar/APISpec.hs
@@ -67,6 +67,7 @@ import qualified SAML2.WebSSO as SAML
 import SAML2.WebSSO.Test.Lenses
 import SAML2.WebSSO.Test.MockResponse
 import SAML2.WebSSO.Test.Util
+import qualified Spar.Data as Data
 import qualified Spar.Intra.Brig as Intra
 import Text.XML.DSig (SignPrivCreds, mkSignCredsWithCert)
 import URI.ByteString.QQ (uri)
@@ -74,6 +75,7 @@ import Util.Core
 import qualified Util.Scim as ScimT
 import Util.Types
 import qualified Web.Cookie as Cky
+import qualified Web.Scim.Schema.User as Scim
 import Wire.API.Cookie
 import Wire.API.Routes.Public.Spar
 import Wire.API.User.IdentityProvider
@@ -746,9 +748,45 @@ specCRUDIdentityProvider = do
         (owner1, _, (^. idpId) -> idpid1, (IdPMetadataValue _ idpmeta1, _)) <- registerTestIdPWithMeta
         (SampleIdP idpmeta2 _ _ _) <- makeSampleIdPMetadata
         _ <- call $ callIdpCreate (env ^. teSpar) (Just owner1) idpmeta2
-        let idpmeta3 = idpmeta1 & edIssuer .~ (idpmeta2 ^. edIssuer)
-        callIdpUpdate' (env ^. teSpar) (Just owner1) idpid1 (IdPMetadataValue (cs $ SAML.encode idpmeta3) undefined)
+        let idpmeta1' = idpmeta1 & edIssuer .~ (idpmeta2 ^. edIssuer)
+        callIdpUpdate' (env ^. teSpar) (Just owner1) idpid1 (IdPMetadataValue (cs $ SAML.encode idpmeta1') undefined)
           `shouldRespondWith` checkErrHspec 400 "idp-issuer-in-use"
+    describe "issuer changed to one that already existed in the same team in the past (but has been updated away)" $ do
+      it "changes back to the old one and keeps the new in the `old_issuers` list." $ do
+        env <- ask
+        (owner1, _, (^. idpId) -> idpid1, (IdPMetadataValue _ idpmeta1, _)) <- registerTestIdPWithMeta
+        idpmeta1' <- do
+          (SampleIdP idpmeta2 _ _ _) <- makeSampleIdPMetadata
+          pure $ idpmeta1 & edIssuer .~ (idpmeta2 ^. edIssuer)
+        idpmeta1'' <- do
+          (SampleIdP idpmeta3 _ _ _) <- makeSampleIdPMetadata
+          pure $ idpmeta1 & edIssuer .~ (idpmeta3 ^. edIssuer)
+
+        do
+          midp <- runSparCass $ Data.getIdPConfig idpid1
+          liftIO $ do
+            (midp ^? _Just . idpMetadata . edIssuer) `shouldBe` Just (idpmeta1 ^. edIssuer)
+            (midp ^? _Just . idpExtraInfo . wiOldIssuers) `shouldBe` Just []
+            (midp ^? _Just . idpExtraInfo . wiReplacedBy) `shouldBe` Just Nothing
+
+        let -- change idp metadata (only issuer, to be precise), and look at new issuer and
+            -- old issuers.
+            change :: SAML.IdPMetadata -> [SAML.IdPMetadata] -> TestSpar ()
+            change new olds = do
+              resp <- call $ callIdpUpdate' (env ^. teSpar) (Just owner1) idpid1 (IdPMetadataValue (cs $ SAML.encode new) undefined)
+              liftIO $ statusCode resp `shouldBe` 200
+
+              midp <- runSparCass $ Data.getIdPConfig idpid1
+              liftIO $ do
+                (midp ^? _Just . idpMetadata . edIssuer) `shouldBe` Just (new ^. edIssuer)
+                sort <$> (midp ^? _Just . idpExtraInfo . wiOldIssuers) `shouldBe` Just (sort $ olds <&> (^. edIssuer))
+                (midp ^? _Just . idpExtraInfo . wiReplacedBy) `shouldBe` Just Nothing
+
+        -- update the name a few times, ending up with the original one.
+        change idpmeta1' [idpmeta1]
+        change idpmeta1'' [idpmeta1, idpmeta1']
+        change idpmeta1 [idpmeta1, idpmeta1', idpmeta1'']
+
     describe "issuer changed to one that is new" $ do
       it "updates old idp, updating both issuer and old_issuers" $ do
         env <- ask
@@ -763,7 +801,7 @@ specCRUDIdentityProvider = do
           statusCode resp `shouldBe` 200
           idp ^. idpMetadata . edIssuer `shouldBe` issuer2
           idp ^. idpExtraInfo . wiOldIssuers `shouldBe` [idpmeta1 ^. edIssuer]
-      it "migrates old users to new idp on their next login" $ do
+      it "migrates old users to new idp on their next login (auto-prov)" $ do
         env <- ask
         (owner1, _, idp1@((^. idpId) -> idpid1), (IdPMetadataValue _ idpmeta1, privkey1)) <- registerTestIdPWithMeta
         issuer2 <- makeIssuer
@@ -780,6 +818,33 @@ specCRUDIdentityProvider = do
         newuref <- tryLogin privkey2 idp2 userSubject
         getUserIdViaRef' olduref >>= \es -> liftIO $ es `shouldBe` Nothing
         getUserIdViaRef' newuref >>= \es -> liftIO $ es `shouldSatisfy` isJust
+
+      it "migrates old users to new idp on their next login (scim)" $ do
+        -- even scim users are automatically updated to a changed IdP issuer.  (this is for
+        -- high-availability; otherwise we could also require the scim peer to push the
+        -- update, and block the old users from logging in until then.)
+        env <- ask
+        (tok, (owner1, _, idp1@((^. idpId) -> idpid1), (IdPMetadataValue _ idpmeta1, privkey1))) <- ScimT.registerIdPAndScimTokenWithMeta
+        issuer2 <- makeIssuer
+        let idpmeta2 = idpmeta1 & edIssuer .~ issuer2
+            privkey2 = privkey1
+            idp2 = idp1 & SAML.idpMetadata .~ idpmeta2
+        let userSubject = SAML.unspecifiedNameID extId
+            extId = "bloob"
+        void $ do
+          -- provision scim user (this is the difference to the above test case)
+          scimusr <- ScimT.randomScimUser
+          ScimT.createUser tok (scimusr {Scim.externalId = Just extId})
+        olduref <- tryLogin privkey1 idp1 userSubject
+        getUserIdViaRef' olduref >>= \es -> liftIO $ es `shouldSatisfy` isJust
+        _ <-
+          let metadata2 = IdPMetadataValue (cs $ SAML.encode idpmeta2) undefined
+           in call $ callIdpUpdate' (env ^. teSpar) (Just owner1) idpid1 metadata2
+        getUserIdViaRef' olduref >>= \es -> liftIO $ es `shouldSatisfy` isJust
+        newuref <- tryLogin privkey2 idp2 userSubject
+        getUserIdViaRef' olduref >>= \es -> liftIO $ es `shouldBe` Nothing
+        getUserIdViaRef' newuref >>= \es -> liftIO $ es `shouldSatisfy` isJust
+
       it "creates non-existent users" $ do
         env <- ask
         (owner1, _, idp1@((^. idpId) -> idpid1), (IdPMetadataValue _ idpmeta1, privkey1)) <- registerTestIdPWithMeta